// src/schemas/public/OrderAddresses.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { UsersId } from './Users';

/** Identifier type for public.order_addresses */
export type OrderAddressesId = number & { __brand: 'OrderAddressesId' };

/** Represents the table public.order_addresses */
export default interface OrderAddresses {
  id: OrderAddressesId;

  user_id: UsersId;

  name: string;

  address: string;

  city: string;

  state: string;

  postal_code: string;

  country: string;
}

/** Represents the initializer for the table public.order_addresses */
export interface OrderAddressesInitializer {
  /** Default value: nextval('order_addresses_id_seq'::regclass) */
  id?: OrderAddressesId;

  user_id: UsersId;

  name: string;

  address: string;

  city: string;

  state: string;

  postal_code: string;

  country: string;
}

/** Represents the mutator for the table public.order_addresses */
export interface OrderAddressesMutator {
  id?: OrderAddressesId;

  user_id?: UsersId;

  name?: string;

  address?: string;

  city?: string;

  state?: string;

  postal_code?: string;

  country?: string;
}