// src/schemas/public/OrderItems.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { OrdersId } from './Orders';
import type { ProductsId } from './Products';

/** Represents the table public.order_items */
export default interface OrderItems {
  order_id: OrdersId;

  product_id: ProductsId;

  quantity: number;

  price_per_unit: number;
}

/** Represents the initializer for the table public.order_items */
export interface OrderItemsInitializer {
  order_id: OrdersId;

  product_id: ProductsId;

  quantity: number;

  price_per_unit: number;
}

/** Represents the mutator for the table public.order_items */
export interface OrderItemsMutator {
  order_id?: OrdersId;

  product_id?: ProductsId;

  quantity?: number;

  price_per_unit?: number;
}